separate analytics form mixin
